Refactored the structure of the game engine a little.
