--- GameEngineJava/src/org/gameEngine/game/GameFactory.java
@@ -1,13 +1,0 @@
-package org.gameEngine.game;
-
-/**
- * Created by TekMaTek on 25/07/2014.
- */
-public class GameFactory {
-
-	public static Game Build( ) {
-		Game newGame = new Game( );
-		return newGame;
-	}
-
-}
